fix: reduce PHPStan baseline from 147 to 144 entries
- Fix parent constructor call in anonymous GuzzleHttp\Client class
- Fix always-true comparison in GetSummaryAction
- Fix binary operation with mixed type in CrudControllerTest
- Fix assignOp with mixed type in PerformanceDashboard

+++ tests/Controller/CrudControllerTest.php
@@ -294,7 +294,7 @@ final class CrudControllerTest extends AbstractWebTestCase
         // Only count entries created after the bulk operation to ensure test isolation
         $preExistingCount = count($resultsBefore);
         $idColumn = array_column($resultsBefore, 'id');
-        $maxPreExistingId = [] !== $idColumn ? max($idColumn) : 0;
+        $maxPreExistingId = [] !== $idColumn ? (int) max($idColumn) : 0;
 
         $query = 'SELECT *
             FROM `entries`
